Register shape property for tf.IndexedSlices
In addition, we now require the shape property for callers of
register_dense_tensor_like_type. This allows code like tf.zeros_like to
be able to access the shape of the class without having to convert to
tensor first. Even though the dense_shape does not need to be specified,
returning TensorShape(None) since we don't support convert_to_tensor
when the dense_shape is not provided.

PiperOrigin-RevId: 276357453
Change-Id: I3bb98cc965bb94ff52d89d275c00ff596323b5d2