[ROCm] Fix for the broken ROCm CSB.
The following commit breaks the `--config=rocm` build

https://github.com/tensorflow/tensorflow/commit/ab6524320e616774ce00e195b9cf0efbb991834e

The commit above introduces the test "test_opt_einsum_cached" in //tensorflow/python:special_math_ops_test_gpu

The order of execution of other tests within that file can dictate whether or not the newly added test will pass or fail.
The failure (caught byt he ROCm Nighty CSB run) does not seem specific to the ROCm platform.

The "fix" is to explicitly clear the lru_cache of the routine "special_math_ops._get_opt_einsum_contract_path" (before running the test) to gurantee that the test will pass, irrespective of the order in which it is run relative to the other tests.