[TF:XLA] Re-implement MatrixSolve in terms of QR decomposition.
This implementation is effectively just a MatrixInverse followed by a batched
dot, so all of the same disclaimers apply as for MatrixInverse (QR decomposition
likely to be slower than LU, doesn't support complex numbers). This is a
short-term solution to avoid having to use workarounds for tf.linalg.solve on
TPU. Performance can be improved in future (e.g. when LU is implemented).

This functionality was added previously but had to be rolled-back due to slow
tests. This version has faster tests.

PiperOrigin-RevId: 277120825
Change-Id: I0a1e274738d9a9393b061f6b793180e1f20dacf7