Fix incorrect RNN gradients with TPUStrategy and mixed precision.
The core issue is that if you call Variable.value() when a TPUStrategy is used, gradients with respect to that Variable are None (I'm not sure why). All the operator overloads in AutoCastVariable used Variable.value(), so they were broken with TPUStrategy. I fixed this by calling Variable.read_value() instead. This fixes any case where an AutoCastVariable operator is used, such as most RNNs.

PiperOrigin-RevId: 277176725
Change-Id: I43e8abcd69f99708d47ec5b3b82b67bab4494db1