Update TF for Bazel 1.0's --incompatible_remove_legacy_whole_archive https://github.com/bazelbuild/bazel/issues/7362
Barring last minute regressions, this is the last commit of this type.

This PR updates `cc_library`s to use `alwayslink=1` when needed. Currently, library archives and library object groups are wrapped in `--whole_archive`, `--no_whole_archive`, thus the whole library is always linked. This changes with Bazel 1.0, and libraries that export otherwise unused symbols must be tagged as `alwayslink=1`.

PiperOrigin-RevId: 277303614
Change-Id: I5b2aefbb9d402bfd1c21f4d2ad95592ba3a392d5