Guarantee no internal references to tf_saved_model exported functions.
The reason for doing this is so that later interprocedural passes don't have
to worry about what to do with these attributes.
An example of where this would happen is when converting to XLA, which
requires eliminating mutable variables. It is thus sort of like an
interprocedural SSA formation, which in particular will
modify signatures interprocedurally. If there are calls to exported
functions, then this conversion cannot cover the entire module.

This also extends the tf_saved_model verifier to check this property.

PiperOrigin-RevId: 277757747
Change-Id: I7b104bdf245dba94c69e7c81e18bb582318fb106