[XLA] Avoid a double conversion in literal_test
This test created a f64 literal from 3.14f.
The closest representable f32 number for 3.14 is 3.1400001049041748046875.
Naturally, 3.1400001049041748046875 is exactly representable in double precision.
However, the closest f64 number for 3.14 is 3.140000000000000124344978758017532527446746826171875.

The intention of this code was to round trip the literal 3.14 so let's just remove the "f" suffix from the code.

Same deal with "2.78f" -> "2.78".

PiperOrigin-RevId: 278416057
Change-Id: Ie2b8a9bb88c93da9913acf3ede70e10c9482dc79