Modular filesystem registration.
This commit contains some additions to the `Env` and `FileSystemRegistry`
classes to support registration of modular filesystem plugins. We also add a new
filesystem class, `ModularFilesystem`, to interface between plugins and core
TensorFlow. At the moment, the class is a stub (to ease reviews and test all
implementations as they get added), but future changes will fill it step by
step.

For details about this, consult the RFC at
https://github.com/tensorflow/community/blob/master/rfcs/20190506-filesystem-plugin-modular-tensorflow.md

PiperOrigin-RevId: 278650824
Change-Id: I97b16f267b4d4caa89ec4cbaab08b8f0f6b8521a