Small targeted optimization to tensorflow::ConvertToEagerTensor.
This is a small optimization that speeds up tf.constant(list of double,
dtype=float64) by 25%, and tf.constant(list of double, dtype=float32) by
about 10%. (In my machines, the benchmarks added in this change go
from 102?s/example to 78?s/example for float64, and from 97?s/example
to 86?s/example for float32). It has no effect on other types of
tf.constant calls.

The optimization consists in:

* Use PyFloat_AS_DOUBLE instead of PyFloat_AsDouble: we know the value
  is valid from the previous PyFloat_Check. This avoids a second
  PyFloat_Check in PyFloat_AsDouble.

* The explicit infinity value assignment is not needed, static_cast<T>
  already does that.

* The overflow check can be simplified to testing whether the casted
  value is infinity while the original is not.

* Lastly, the overflow check only needs to be done if sizeof(T) <
  sizeof(double). Since the function is a template in T, the entire
  overflow check branch is optimized away by the compiler when T == double.

PiperOrigin-RevId: 278888241
Change-Id: I6632201fc3f9bef5f7bd2eadecb7280fd851c6d8