TFLM: online planning PR 1 of many: Allocate node and reg from allocator.
To support scratch buffers, we need to keep the state of nodes for the entire lifetime of the model because:
- the scratch buffer tensor shapes needs to be available and can't be overwritten
- the value for persistent tensors needs to be alive forever.

This will slightly increase the arena size required, but with recent CLs which reduced unnecessary data entries, developer won't need to increase the arena size hopefully.

This will also make the code run faster as it doesn't need to reconstruct node as well as running `prepare` and `init` method for each invocation.

PiperOrigin-RevId: 279181728
Change-Id: Icc35cfd43288e47e720cfba60d726494bab08353