Add new_writable_file support to the modular POSIX filesystem.
Also add tests for this API entry-point to ensure API requirements is satisfied and implementation creates the file. One test requires create_directory but that implementation would come in a subsequent CL. Hence, we add helper test method for the case when an API entry point that is needed to create test setup is not implemented by the plugin under test.

Part of work for modular filesystem plugins. For more details, consult the RFC at
https://github.com/tensorflow/community/blob/master/rfcs/20190506-filesystem-plugin-modular-tensorflow.md

PiperOrigin-RevId: 279762912
Change-Id: I66624dee4bfe46db5a2fc34953b60db9e617b777